requirements.txt
django==2.2.16
mysqlclient
django-debug-toolbar
pillow==6.2.0
djangorestframework
bcrypt
django-filter
djangorestframework-jwt
pymysql
django-cors-headers
gevent
coreapi
python-memcached
openpyxl
drf-yasg
django-redis
redis
celery